API + STREAMLIT/requirements.txt
fastapi
uvicorn
google-cloud-storage
pandas
numpy==1.23.5
streamlit
altair==4.2.0
matplotlib
google-cloud-bigquery
pyyaml
prophet
scikit-learn==1.3.2 